Install node-schedule + time TTL for twits
In stead of using the MongoDB Index for TTL deletion, I used the npm pakcage "node-schedule" for deletion. It is way more precise and can be managed from the meteor code base. It now works till the theoretical nanosecond.

Add another function for increasing the TTL time, but this one isn't tested yet. Deletion works.